harvester/requirements.txt
astroid==2.3.3
autopep8==1.5.1
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
pycodestyle==2.5.0
pylint==2.4.4
PySocks==1.7.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
tweepy==3.8.0
typed-ast==1.4.1
urllib3==1.25.8
wrapt==1.11.2
